requirements_old.txt
numpy==1.26.4
torch==2.4.1
transformers==4.44.1
flask~=3.0.3
werkzeug~=3.0.3
matplotlib~=3.9.2
pillow~=10.4.0
#diffusers~=0.18.2